requirements.txt
# 核心框架
fastapi>=0.104.0
uvicorn>=0.23.0
langchain>=0.1.0
langchain-openai>=0.1.0
langchain-community>=0.1.0

# 向量数据库
faiss-cpu>=1.7.4
chromadb>=0.4.0

# 数据处理
pandas>=2.0.0
PyPDF2>=3.0.1
pdfminer.six>=20221105
beautifulsoup4>=4.12.0
requests>=2.31.0

# 大语言模型相关
openai>=1.3.0
sentence-transformers>=2.2.0
transformers>=4.35.0
torch>=2.0.0

# 知识图谱
networkx>=3.1
neo4j>=5.14.0

# 情感分析
snownlp>=0.12.3

# 缓存
redis>=5.0.0

# 前端界面
streamlit>=1.28.0

# 工具类
python-dotenv>=1.0.0
pydantic>=2.0.0

# 异步支持
asyncio
aiofiles>=23.2.0

# 并发处理
concurrent-log-handler>=0.9.25